mixcontent/requirements.txt
Django==2.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
django-rest-swagger==2.2.0
PyJWT==1.6.4
pytz==2018.5
psycopg2=2.7.5
